feat(ResultView): ResultView 클래스를 프로덕션으로 이동
