+++ src/test/java/org/baseball/domain/ResultViewTest.java
@@ -40,72 +40,12 @@ class ResultViewTest {
         assertThatCode(() -> resultView.exit())
                 .doesNotThrowAnyException();
     }
-
-
+    
     private BallStatus setBallStatus(int strikeCount) {
         ballStatus = new BallStatus(3);
         for (int i = 0; i < strikeCount; i++) {
             ballStatus.increaseStrikeCount();
         }
         return ballStatus;
-    }
-}
-
-class ResultView {
-    public GameStatus conclude(BallStatus ballStatus) {
-        return showResult(ballStatus);
-    }
-
-    public void exit() {
-        end();
-    }
-
-    private GameStatus showResult(BallStatus ballStatus) {
-        if (ballStatus.isNothing()) {
-            nothing();
-            return GameStatus.FAIL;
-        }
-
-        String result = buildResult(ballStatus);
-        System.out.println(result);
-
-        if (ballStatus.allStrike()) {
-            gameClear();
-            restart();
-            return GameStatus.CLEAR;
-        }
-
-        return GameStatus.FAIL;
-    }
-
-    private void nothing() {
-        System.out.println(ConsoleMessage.NOTHING);
-    }
-
-    private String buildResult(BallStatus ballStatus) {
-        StringBuilder sb = new StringBuilder();
-
-        sb.append(ballStatus.getBallCount())
-                .append(ConsoleMessage.BALL);
-
-        if (ballStatus.hasStrikeCount()) {
-            sb.append(" ")
-                    .append(ballStatus.getStrikeCount())
-                    .append(ConsoleMessage.STRIKE);
-        }
-
-        return sb.toString();
-    }
-
-    private void gameClear() {
-        System.out.println(ConsoleMessage.GAME_CLEAR);
-    }
-
-    private void restart() {
-        System.out.println(ConsoleMessage.GAME_RESTART);
-    }
-
-    private void end() {
-        System.out.println(ConsoleMessage.GAME_END);
     }
 }
